Windows: Address an VS2012/x64 warning in hotplug.c
..\libusb\hotplug.c(255): warning C4244: '=' : conversion from
'ssize_t' to 'int', possible loss of data

+++ libusb/hotplug.c
@@ -252,7 +252,7 @@ int API_EXPORTED libusb_hotplug_register_callback(libusb_context *ctx,
 		int i, len;
 		struct libusb_device **devs;
 
-		len = libusb_get_device_list(ctx, &devs);
+		len = (int) libusb_get_device_list(ctx, &devs);
 		if (len < 0) {
 			libusb_hotplug_deregister_callback(ctx,
 							new_callback->handle);
